+++ include/mujoco/mjmodel.h
@@ -58,8 +58,9 @@ typedef enum mjtDisableBit_ {     // disable default feature bitflags
   mjDSBL_FILTERPARENT = 1<<9,     // remove collisions with parent body
   mjDSBL_ACTUATION    = 1<<10,    // apply actuation forces
   mjDSBL_REFSAFE      = 1<<11,    // integrator safety: make ref[0]>=2*timestep
+  mjDSBL_SENSOR       = 1<<12,    // sensors
 
-  mjNDISABLE          = 12        // number of disable flags
+  mjNDISABLE          = 13        // number of disable flags
 } mjtDisableBit;